cpufreq: conservative: Fix compile error

+++ drivers/cpufreq/cpufreq_conservative.c
@@ -94,7 +94,7 @@ static struct dbs_tuners {
 	unsigned int ignore_nice;
 	unsigned int freq_step;
 	unsigned int input_boost_freq;
-	u64 input_boost_duration;
+	unsigned int input_boost_duration;
 } dbs_tuners_ins = {
 	.up_threshold = DEF_FREQUENCY_UP_THRESHOLD,
 	.down_threshold = DEF_FREQUENCY_DOWN_THRESHOLD,
